+++ examples/lib/db.dart
@@ -5,7 +5,7 @@ import 'models/todo.dart';
 import 'models/user.dart';
 
 final appDb = DB.autoVersion(
-  databaseName: 'sqflow_showcase.db',
+  databaseName: 'phorm_showcase.db',
   tables: [
     usersTable,
     postsTable,