+++ examples/test/widget_test.dart
@@ -7,10 +7,10 @@ import 'package:phorm_example/pages/social_feed_page.dart';
 import 'package:phorm_example/pages/validation_demo_page.dart';
 
 void main() {
-  testWidgets('SqflowShowcaseApp navigation smoke test',
+  testWidgets('PhormShowcaseApp navigation smoke test',
       (WidgetTester tester) async {
     // Build our app and trigger a frame.
-    await tester.pumpWidget(const SqflowShowcaseApp());
+    await tester.pumpWidget(const PhormShowcaseApp());
 
     // Verify that the NavigationBar is present.
     expect(find.byType(NavigationBar), findsOneWidget);