+++ phorm_generator/CHANGELOG.md
@@ -13,5 +13,5 @@
 - First stable release.
 - SQL schema generation for Flutter based on annotations.
 - Integration with `build_runner` and `source_gen` for automatic code generation.
-- Support for the `sqflow_annotations` package.
+- Support for the `phorm_annotations` package.
 - Automatic formatting of generated Dart code using `dart_style`.